moved *group.unwrap() from core.topologyattrs to core.groups
